Add top menu to video detail screen
